feat(web-ui): add dedicated Scoped Models settings page
- Create new /settings/scoped-models route with full page for scoped model management
- Add Scoped Models to settings dropdown (nav-secondary.tsx)
- Add Scoped Models card to settings index page
- Remove embedded ScopedModelsSection from Auth settings page
- Update nav-secondary tests to include Scoped Models link
- Regenerate route tree for new route

Fixes #183
